requirements.txt
anyio==3.5.0
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
fastapi==0.73.0
flake8==4.0.1
h11==0.13.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
python-dotenv==0.19.2
requests==2.27.1
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
tomli==2.0.0
typing-extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.1
